+++ public/components/common/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './datasource_utils';
+export { DataSourceSelector } from './datasource_selector';
+export { DeleteModal } from './DeleteModal';
+export { DashboardInstallModal } from './dashboard_install_modal';