+++ public/components/experiment/services/experiment_service.ts
@@ -10,25 +10,34 @@ import { combineResults, toExperiment, toExperimentSchedule } from '../../../typ
 export class ExperimentService {
   constructor(private http: CoreStart['http']) {}
 
-  async getExperiments() {
-    const response = await this.http.get(ServiceEndpoints.Experiments);
+  async getExperiments(dataSourceId?: string) {
+    const query = dataSourceId ? { dataSourceId } : undefined;
+    const response = await this.http.get(ServiceEndpoints.Experiments, { query });
     return combineResults(
       ...(response ? response.hits.hits.map((hit) => toExperiment(hit._source)) : [])
     );
   }
 
-  async getExperiment(id: string) {
-    return await this.http.get(`${ServiceEndpoints.Experiments}/${id}`);
+  async getExperiment(id: string, dataSourceId?: string) {
+    const query = dataSourceId ? { dataSourceId } : undefined;
+    return await this.http.get(`${ServiceEndpoints.Experiments}/${id}`, { query });
   }
 
-  async createExperiment(data: any) {
+  async createExperiment(data: any, datasourceId?: string) {
+    const query = datasourceId ? { dataSourceId: datasourceId } : undefined;
+    
     return await this.http.post(ServiceEndpoints.Experiments, {
       body: JSON.stringify(data),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      ...(query && { query }),
     });
   }
 
-  async deleteExperiment(id: string) {
-    return await this.http.delete(`${ServiceEndpoints.Experiments}/${id}`);
+  async deleteExperiment(id: string, dataSourceId?: string) {
+    const query = dataSourceId ? { dataSourceId } : {};
+    return await this.http.delete(`${ServiceEndpoints.Experiments}/${id}`, { query });
   }
 
   async getScheduledExperiments() {
@@ -49,7 +58,8 @@ export class ExperimentService {
     });
   }
 
-  async deleteScheduledExperiment(id: string) {
-    return await this.http.delete(`${ServiceEndpoints.ScheduledExperiments}/${id}`);
+  async deleteScheduledExperiment(id: string, dataSourceId?: string) {
+    const query = dataSourceId ? { dataSourceId } : {};
+    return await this.http.delete(`${ServiceEndpoints.ScheduledExperiments}/${id}`, { query });
   }
 }