+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Features
 
 - Add Search Pipeline to Search Configuration Detail page ([#699](https://github.com/opensearch-project/dashboards-search-relevance/issues/699))
+- Support multiple datasource in SRW ([#723](https://github.com/opensearch-project/dashboards-search-relevance/pull/723))
 
 ### Enhancements